topic 6180: fix restoring of some geometry classes
git-svn-id: http://jmonkeyengine.googlecode.com/svn/trunk@3820 75d07b2b-3a1a-0410-a2c5-0572b91ccdca

+++ src/com/jme/scene/shape/Dodecahedron.java
@@ -52,7 +52,7 @@ import com.jme.util.geom.BufferUtils;
  * <code>Dodecahedron</code>
  *
  * @author Joshua Slack
- * @version $Revision: 1.2 $
+ * @version $Revision: 1.3 $
  *
  */
 public class Dodecahedron extends TriMesh {
@@ -220,7 +220,7 @@ public class Dodecahedron extends TriMesh {
     public void read(JMEImporter e) throws IOException {
         super.read(e);
         InputCapsule capsule = e.getCapsule(this);
-        sideLength = capsule.readInt("sideLength", 0);
+        sideLength = capsule.readFloat("sideLength", 0);
         
     }
